+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/order/service/OrderApproveService.java
@@ -15,7 +15,7 @@ public class OrderApproveService {
 
     private final OrderAdaptor orderAdaptor;
 
-    @RedissonLock(LockName = "주문승인", identifier = "orderUuid")
+    @RedissonLock(LockName = "주문", identifier = "orderUuid")
     public String execute(String orderUuid, Long userId) {
         Order order = orderAdaptor.findByOrderUuid(orderUuid);
         // TODO : 승인 주문 관리자 할당.
